add a test, update 1 more type

+++ graphql_api/types/repository_config/repository_config.py
@@ -17,13 +17,15 @@ class IndicationRange(TypedDict):
 
 
 @repository_config_bindable.field("indicationRange")
-async def resolve_indication_range(repository: Repository, info) -> dict[str, int]:
+async def resolve_indication_range(repository: Repository, info) -> dict[str, float]:
     owner = await OwnerLoader.loader(info).load(repository.author_id)
 
     yaml = await sync_to_async(UserYaml.get_final_yaml)(
         owner_yaml=owner.yaml, repo_yaml=repository.yaml
     )
-    range: list[int] = yaml.get("coverage", {"range": [60, 80]}).get("range", [60, 80])
+    range: list[float] = yaml.get("coverage", {"range": [60, 80]}).get(
+        "range", [60, 80]
+    )
     return {"lowerRange": range[0], "upperRange": range[1]}
